9624: Move test files next to components being tested
