+++ app/javascript/components/Filters/testUtils.js
@@ -1,6 +1,6 @@
-import FiltersModel from '../../../../app/javascript/components/Filters/model';
-import ConditionSetModel from '../../../../app/javascript/components/ConditionSetFormField/model';
-import { submitterType } from '../../../../app/javascript/components/Filters/SubmitterFilter/component';
+import FiltersModel from './model';
+import ConditionSetModel from '../ConditionSetFormField/model';
+import { submitterType } from './SubmitterFilter/component';
 
 export const getFiltersStore = () => new FiltersModel({
   conditionSetStore: new ConditionSetModel({
